modified the adapter for native
