+++ test/spec/modules/precisoBidAdapter_spec.js
@@ -1,5 +1,8 @@
 import { expect } from 'chai';
 import { spec } from '../../../modules/precisoBidAdapter.js';
+import { NATIVE } from '../../../src/mediaTypes.js';
+import { OPENRTB } from '../../../libraries/precisoUtils/bidUtils.js';
+
 // simport { config } from '../../../src/config.js';
 
 const DEFAULT_PRICE = 1
@@ -53,6 +56,74 @@ describe('PrecisoAdapter', function () {
 
   };
 
+  let nativeBid = {
+
+    precisoBid: true,
+    bidId: '23fhj33i987f',
+    bidder: 'precisonat',
+    buyerUid: 'testuid',
+    params: {
+      host: 'prebid',
+      sourceid: '0',
+      publisherId: '0',
+      mediaType: 'native',
+      region: 'IND'
+
+    },
+    userId: {
+      pubcid: '12355454test'
+
+    },
+    user: {
+      geo: {
+        region: 'IND',
+      }
+    },
+    ortb2: {
+      device: {
+        w: 1920,
+        h: 166,
+        dnt: 0,
+      },
+      site: {
+        domain: 'localHost'
+      },
+      source: {
+        tid: 'eaff09b-a1ab-4ec6-a81e-c5a75bc1dde3'
+      }
+
+    },
+    mediaType: 'native',
+    nativeOrtbRequest: {
+      assets: [
+        {
+          id: 3,
+          required: 1,
+          img: {
+            type: 3,
+            w: 300,
+            h: 250
+          }
+        }
+      ]
+    },
+    nativeParams: {
+      ortb: {
+        assets: [
+          {
+            id: 3,
+            required: 1,
+            img: {
+              type: 3,
+              w: 300,
+              h: 250
+            }
+          }
+        ]
+      }
+    }
+  };
+
   describe('isBidRequestValid', function () {
     it('Should return true if there are bidId, params and sourceid parameters present', function () {
       expect(spec.isBidRequestValid(bid)).to.be.true;
@@ -60,6 +131,14 @@ describe('PrecisoAdapter', function () {
     it('Should return false if at least one of parameters is not present', function () {
       delete bid.params.publisherId;
       expect(spec.isBidRequestValid(bid)).to.be.false;
+    });
+    it('Should return true if there are bidId, params and sourceid parameters present native Bid', function () {
+      expect(spec.isBidRequestValid(nativeBid)).to.be.true;
+    });
+
+    it('Should return false if at least one of parameters is not present in native bid', function () {
+      delete nativeBid.params.publisherId;
+      expect(spec.isBidRequestValid(nativeBid)).to.be.false;
     });
   });
 
@@ -90,6 +169,31 @@ describe('PrecisoAdapter', function () {
       serverRequest = spec.buildRequests([bid]);
       let data = serverRequest.data;
       expect(data.device).to.be.undefined;
+    });
+
+    let ServeNativeRequest = spec.buildRequests([nativeBid]);
+    it('Creates a valid nativeServerRequest object ', function () {
+      expect(ServeNativeRequest).to.exist;
+      expect(ServeNativeRequest.method).to.exist;
+      expect(ServeNativeRequest.url).to.exist;
+      expect(ServeNativeRequest.data).to.exist;
+      expect(ServeNativeRequest.method).to.equal('POST');
+      expect(ServeNativeRequest.url).to.equal('https://ssp-bidder.mndtrk.com/bid_request/openrtb');
+    });
+
+    it('should extract the native params', function () {
+      let nativeData = ServeNativeRequest.data;
+      const asset = JSON.parse(nativeData.imp[0].native.request).assets[0]
+      expect(asset).to.deep.equal({
+        id: OPENRTB.NATIVE.ASSET_ID.IMAGE,
+        required: 1,
+        img: {
+          w: 300,
+          h: 250,
+          type: OPENRTB.NATIVE.IMAGE_TYPE.MAIN,
+        }
+      }
+      )
     });
   });
 
@@ -137,7 +241,87 @@ describe('PrecisoAdapter', function () {
 
       expect(Object.keys(result[0])).to.have.members(Object.keys(expectedResponse[0]))
     })
+
+    it('should get correct native bid response', function () {
+      const adm = {
+        native: {
+          ver: 1.2,
+          link: {
+            url: 'https://example.com',
+            clicktrackers: 'https://example.com/clktracker'
+          },
+          eventtrackers: [
+            {
+              url: 'https://example.com/imptracker'
+            }
+          ],
+          imptrackers: [
+            'https://example.com/imptracker'
+          ],
+          assets: [{
+            id: OPENRTB.NATIVE.ASSET_ID.IMAGE,
+            required: 1,
+            img: {
+              url: 'https://example.com/image.jpg',
+              w: 150,
+              h: 50
+            }
+          }],
+        }
+      }
+      let nativeResponse = {
+        bidderRequestId: 'f6adb85f-4e19-45a0-b41e-2a5b9a48f23a',
+        seatbid: [
+          {
+            bid: [
+              {
+                id: '123',
+                impid: 'b4f290d7-d4ab-4778-ab94-2baf06420b22',
+                price: DEFAULT_PRICE,
+                adm: JSON.stringify(adm),
+                cid: 'test_cid',
+                crid: 'test_banner_crid',
+                w: DEFAULT_BANNER_WIDTH,
+                h: DEFAULT_BANNER_HEIGHT,
+                adomain: [],
+              }
+            ],
+            seat: BIDDER_CODE
+          }
+        ],
+      }
+
+      let expectedNativeResponse = [
+        {
+          requestId: 'b4f290d7-d4ab-4778-ab94-2baf06420b22',
+          mediaType: NATIVE,
+          cpm: DEFAULT_PRICE,
+          creativeId: 'test_banner_crid',
+          width: 1,
+          height: 1,
+          ttl: 300,
+          meta: {
+            advertiserDomains: []
+          },
+          netRevenue: true,
+          currency: 'USD',
+          // meta: { advertiserDomains: [] },
+          native: {
+            clickUrl: encodeURI('https://example.com'),
+            impressionTrackers: ['https://example.com/imptracker'],
+            image: {
+              url: encodeURI('https://example.com/image.jpg'),
+              width: 150,
+              height: 50
+            },
+          }
+        }
+      ]
+      let result = spec.interpretResponse({ body: nativeResponse });
+      expect(Object.keys(result[0])).to.have.members(Object.keys(expectedNativeResponse[0]));
+    })
   })
+
   describe('getUserSyncs', function () {
     const syncUrl = 'https://ck.2trk.info/rtb/user/usersync.aspx?id=NA&gdpr=0&gdpr_consent=&us_privacy=&t=4';
     const syncOptions = {
